Drop support for cpy version of snakeoil._klass in preparation for cython version.
The intent here is to just use cython annotations against the native python
implementations to gain performance gains.

This should remove the need to maintain duplicate implementations, while
giving the desired optimization gains for critical/core pathways in klass.

Note: a pure .pyx could be written that would probably be a bit faster- same
for writing raw cpy.  That's a followup optimization if warranted.